@@ -6,12 +6,11 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@mui/icons-material": "^5.11.0",
-    "@mui/material": "^5.11.5",
-    "@mui/styles": "^5.11.2",
+    "@mui/material": "^5.11.8",
     "mapbox-gl": "2.12.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "6.4.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6.8.1",
     "react-scripts": "5.0.1"
   },
   "devDependencies": {
